Server/marrir-apiZ/pyproject.toml
[project]
name = "marrir-apix"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "apscheduler>=3.11.0",
    "authlib>=1.5.1",
    "bcrypt==4.0.1",
    "boto3>=1.37.4",
    "docuseal>=1.0.4",
    "faker>=36.1.1",
    "fastapi[standard]>=0.115.10",
    "google-api-core>=2.24.1",
    "google-auth>=2.38.0",
    "google-cloud-storage>=3.1.0",
    "google-crc32c>=1.6.0",
    "google-resumable-media>=2.7.2",
    "itsdangerous>=2.2.0",
    "opencv-python>=4.11.0.86",
    "opencv-python-headless>=4.11.0.86",
    "openpyxl>=3.1.5",
    "pandas>=2.2.3",
    "passlib[bcrypt]>=1.7.4",
    "pdfkit>=1.0.0",
    "phonenumbers>=8.13.55",
    "playwright>=1.51.0",
    "psycopg2-binary>=2.9.10",
    "pydantic-extra-types>=2.10.2",
    "pydantic-settings>=2.8.1",
    "python-jose[cryptography]>=3.4.0",
    "qrcode>=8.0",
    "readmrz>=0.0.2",
    "sqlalchemy>=2.0.38",
    "stripe>=11.6.0",
    "telr-payment>=0.1.3.1",
    "weasyprint>=65.0",
]
